requirements.txt
# 智览系统依赖包
pyyaml>=6.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0
networkx>=3.1
jieba>=0.42.1
nltk>=3.8.1
jinja2>=3.1.2
pylatex>=1.4.1
colorlog>=6.7.0
openai>=1.0.0
dashscope>=1.14.0
tqdm>=4.66.0
python-dateutil>=2.8.0
pydantic>=2.0.0
